requirements.txt
numpy==1.26.4
opencv-python==4.10.0.82
pillow
scikit-learn==1.5.0
scipy==1.13.1
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
